+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.92
 
+- Amazon: removed hover on scroll, since it is annoying
 - fixed weird formatting of sliders
 - fix: language button